requirements.txt
python-dotenv==0.19.2
Flask==2.0.3
requests==2.20.0
ibm_watson_machine_learning==1.0.204
Flask-WTF==1.0.0
